Reorganise fmm module
