+++ fmm/Cargo.toml
@@ -30,17 +30,10 @@ cauchy = "0.4.*"
 itertools = "0.10"
 mpi = { version = "0.6.*" }
 rand = "0.8.*"
-ndarray = { version = "*", features = ["blas"]}
-ndarray-rand = "0.14.0"
-blas-src = { version = "*", features = ["openblas"] }
-openblas-src = { version = "*", features = ["cblas", "system"] }
-ndarray-linalg = { version = "*", features = ["openblas-system"] }
 float-cmp = "0.9.0"
 rayon = "1.7"
 num_cpus = "1"
-ndrustfft = "0.4.0"
 num = "0.4"
-ndarray-ndimage = "0.3.0"
 rlst = {git = "https://github.com/skailasa/rlst.git", branch = "enh/moore-penrose-pseudo-inverse" }
 
 [target.aarch64-apple-darwin]
