+++ fmm/src/lib.rs
@@ -2,4 +2,6 @@
 #![cfg_attr(feature = "strict", deny(warnings))]
 pub mod charge;
 pub mod fmm;
-pub mod impl_charge;
+pub mod types;
+pub mod field_translation;
+pub mod interaction_lists;
